pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lampson"
version = "0.1.0"
description = "自更新CLI智能助手"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",
    "prompt_toolkit>=3.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lark-oapi>=1.5.0",
]

[project.scripts]
lampson = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
